requirements.txt
numpy==2.2.3
opencv_contrib_python==4.11.0.86
Pillow==11.1.0
pyttsx3==2.98
